requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.47.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
streamlit==1.47.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
